% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_columnPair.R
\name{add_columnPair}
\alias{add_columnPair}
\title{Adding the column "pair" to list of data frames}
\usage{
add_columnPair(cellDist)
}
\arguments{
\item{cellDist}{List of data frames where each data frame represents one time point 
and has the following columns: names from cell population 1, names from cell population 2 and distance.}
}
\value{
List of data frames as above with one additional column "pair" 
where the names of the cell pairs are put into one string in the format "Track_[id1]-Track[id2]".
}
\description{
Adding the column "pair" to list of data frames
}
