% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_meas_to_matches.R
\name{add_meas_to_matches}
\alias{add_meas_to_matches}
\title{Adding end point measurements to matches of tracks tp endpoint ROIs}
\usage{
add_meas_to_matches(matches, endPointMeasTxt)
}
\arguments{
\item{matches}{Data frame of pairs of tracked cell names and endpoint cells which have been matched to each other.}

\item{endPointMeasTxt}{Data frame with measurement results of endpoint ROIs, for instance from Fiji, loaded as .csv,}
}
\value{
Data frame with the following columns:
ROI names from the endpoint, from the imaging; the area, mean, min and max intensity of the endpoint ROI.
}
\description{
Adding end point measurements to matches of tracks tp endpoint ROIs
}
