% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_contacts.R
\name{analyze_contacts}
\alias{analyze_contacts}
\title{Compute cell contact duration and number of contacts}
\usage{
analyze_contacts(contactList, secPerFrame = Inf)
}
\arguments{
\item{contactList}{List of data frames. Each data frame represents one cell-cell contacts
with the columns "pair", "distance", "startTime", "endTime", "duration", "tumorCell", "TCell".}

\item{secPerFrame}{Time between each frame in seconds.}
}
\value{
A data frame with the columns "tumorCell" and the cumulative duration and number of contacts.
}
\description{
Compute cell contact duration and number of contacts
}
