% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_distTimepoint_withoutGrid.R
\name{compute_distTimepoint_withoutGrid}
\alias{compute_distTimepoint_withoutGrid}
\title{Title: Computing cell-cell distances for one time point}
\usage{
compute_distTimepoint_withoutGrid(
  timepoint,
  gridWidth = 100,
  ROImap1 = tumourROImap,
  ROImap2 = tcellROImap,
  ROImapField1 = tumourROImap_NamesToGrid,
  ROImapField2 = tcellROImap_NamesToGrid
)
}
\arguments{
\item{timepoint}{An integer that indicates the time point for which cell-cell
distances should be computed.}

\item{gridWidth}{The width of one square in the grid into which ROI centers have been assigned.}

\item{ROImap1}{Hash map with tumour ROI names as keys and tumour ROI coordinates as values.}

\item{ROImap2}{Hash map with T cell ROI names as keys and T cell ROI coordinates as values.}

\item{ROImapField1}{Hash map with tumour ROI names as keys and assignment to grid as value.}

\item{ROImapField2}{Hash map with T cell ROI names as keys and assignment to grid as value.}
}
\value{
A data frame with the cell-cell distances of all possible cell pairs at the chosen time points.
If the cell-cell distance is larger than the gridWidth, ">>[gridWidth]" will be inserted as value.
}
\description{
For all possible cell pairs in one frame, this function checks whether the center
points of the cell pairs are in at least neighboring squares of the 2D grid.
If yes, the function computes the cell-cell distances.
}
