% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid.R
\name{create_grid}
\alias{create_grid}
\title{Helper function to create a 2D grid}
\usage{
create_grid(xlim, ylim, dim = 5)
}
\arguments{
\item{xlim}{x-dimensions of the grid in the format "c([start], [end])"}

\item{ylim}{y-dimensions of the grid in the format "c([start], [end])"}

\item{dim}{Number of rows/columns in the grid.}
}
\value{
Data frame representing the cells of a grid with the selected dimensions.
}
\description{
Helper function to create a 2D grid
}
\details{
The grid should be slightly bigger than the dimensions of the frame 
as cell centers could lay outside the grid.
}
