% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pairs.R
\name{get_pairs}
\alias{get_pairs}
\title{Filtering data frame with cell-cell distances for cell-cell contacts}
\usage{
get_pairs(x, minimumDuration = 5)
}
\arguments{
\item{x}{A data frame containing the columns "pair", "distance" and "contact". 
The latter should be a column with TRUE/FALSE entries. The entries of the "pair"
column should be in the formal "[name of tumor cell]-[name of T cell]"}

\item{minimumDuration}{The minimum duration of a contact}
}
\value{
A data frame containing the columns "pair", "distance", "startTime", "endTime", "duration", "tumorCell", "TCell"
}
\description{
Filtering data frame with cell-cell distances for cell-cell contacts
}
