% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pairs_wROInames.R
\name{get_pairs_wROInames}
\alias{get_pairs_wROInames}
\title{Title}
\usage{
get_pairs_wROInames(x, minimumDuration = 5)
}
\arguments{
\item{x}{Data frame which contains cell-cell distances below distance threshold with columns pair, distance, and time point}

\item{minimumDuration}{Minimum duration of a 'cell contact'}
}
\value{
Data frame which contains only cell-cell distances which lasts for the minimum duration or longer
}
\description{
Title
}
