% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_Hashmap_NamesToCoords.R
\name{insert_Hashmap_NamesToCoords}
\alias{insert_Hashmap_NamesToCoords}
\title{Insert ROI names as keys and ROI coordinates as values into hash map}
\usage{
insert_Hashmap_NamesToCoords(ROIs, ROImap)
}
\arguments{
\item{ROIs}{List of ROIs, usually loaded from a .zip file via the "read.ijzip" function.}

\item{ROImap}{A hash map into which key and values should be inserted.}
}
\value{
There is no return value but the hash map will be edited.
}
\description{
Insert ROI names as keys and ROI coordinates as values into hash map
}
