% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_Hashmap_NamesToGrid.R
\name{insert_Hashmap_NamesToGrid}
\alias{insert_Hashmap_NamesToGrid}
\title{Insert ROI names as keys and assignment to cells in grid as values into hash map}
\usage{
insert_Hashmap_NamesToGrid(ROIs, ROImap, grid)
}
\arguments{
\item{ROIs}{List of ROIs, usually loaded from a .zip file via the "read.ijzip" function.}

\item{ROImap}{A hash map into which key and values should be inserted.}

\item{grid}{Data frame representing a grid. ROI centers are assigned into cells of this grid.}
}
\value{
There is no return value but the hash map will be edited.
}
\description{
Insert ROI names as keys and assignment to cells in grid as values into hash map
}
