% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_contactTrue.R
\name{keep_contactTrue}
\alias{keep_contactTrue}
\title{Keeping only cell pairs with distance below threshold.}
\usage{
keep_contactTrue(cellDistances)
}
\arguments{
\item{cellDistances}{Data frame with column "distances"}
}
\value{
Data frame as input but with additional column "contact" with TRUE/FALSE entries.
}
\description{
Keeping only cell pairs with distance below threshold.
}
