% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_to_endpoint_ROIs.R
\name{match_to_endpoint_ROIs}
\alias{match_to_endpoint_ROIs}
\title{Match tracked tumour cells to endpoint tumour cells}
\usage{
match_to_endpoint_ROIs(
  endPointROIs,
  endpointROImap,
  tumorROImap,
  tumorROImap_center,
  tumorROImap_NamesToGrid
)
}
\arguments{
\item{endPointROIs}{List of ROIs at the last time point, usually loaded from a .zip file via the "read.ijzip" function.}

\item{endpointROImap}{Hash map of endpoint ROI names as keys and endpoint ROI coordinates as values.}

\item{tumorROImap}{Hash map of names of tracked ROIs as keys and their coordinates as values.}

\item{tumorROImap_center}{Hash map of names of tracked ROIs as keys and their center as values.}

\item{tumorROImap_NamesToGrid}{Hash map of names of tracked ROIs as keys and their assignment to grid as values.}
}
\value{
Data frame of pairs of tracked cell names and endpoint cells which have been matched to each other.
}
\description{
Match tracked tumour cells to endpoint tumour cells
}
