% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_matches.R
\name{plot_matches}
\alias{plot_matches}
\title{Plotting matches of tracked ROIs to end point ROIs}
\usage{
plot_matches(
  matches,
  tumorROImap,
  endpointROImap,
  cols = brewer.pal(10, "Spectral"),
  xlim = c(0, 550),
  ylim = c(0, 550)
)
}
\arguments{
\item{matches}{Data frame with ROI name from tracks in one column and 
matched tumour ROI name from endpoint in second column.}

\item{tumorROImap}{Hash map with tracked ROI names as keys and ROI coordinates as values.}

\item{endpointROImap}{Hash map with endpoint ROI names as keys and ROI coordinates as values.}

\item{cols}{Vector with colors. The default is set to "brewer.pal( 10,'Spectral')".}

\item{xlim}{Width of the frame.}

\item{ylim}{Hight of the frame.}
}
\value{
There is no return value but a plot will be displayed.
}
\description{
Plotting matches of tracked ROIs to end point ROIs
}
