% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_export_contacts_for_df_wROInames.R
\name{prepare_export_contacts_for_df_wROInames}
\alias{prepare_export_contacts_for_df_wROInames}
\title{Prepare computed contacts for export as .csv file}
\usage{
prepare_export_contacts_for_df_wROInames(
  pairs,
  tracksTxt,
  cellType = "tumourCell"
)
}
\arguments{
\item{pairs}{Cell pairs in contact}

\item{tracksTxt}{Tracks file with Track ID, frame number, xy-position}

\item{cellType}{Celltype can either by "tumourCell" or "Tcell" (use this parameter to add it as variable in the exported df)}
}
\description{
Prepare computed contacts for export as .csv file
}
