% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareExportContacts.R
\name{prepareExportContacts}
\alias{prepareExportContacts}
\title{Creating data frame with tumor _or_ T cells in contact}
\usage{
prepareExportContacts(pairs, tracksTxt, cellType = "tumorCell")
}
\arguments{
\item{pairs}{List of computed cell pairs that were in contact 
with column "Var1" for tumor cells and "Var2" for T cells.}

\item{tracksTxt}{Data frame of exported tracks. Usually, 
the data was previously exported from TrackMate or Imaris and loaded as table from a .csv file.}

\item{cellType}{Either "tumorCell" or "TCell"}
}
\value{
The data frame of exported tracks reduced to only the rows that represent
cell ROIs that were involved in a contact.
}
\description{
Creating data frame with tumor _or_ T cells in contact
}
